Final Deliverable/Final Code - Inventory Managment System for Retailers/Backend/sendgrid.py
import smtplib
import os
from dotenv import load_dotenv
from email.mime.multipart import MIMEMultipart
from email.mime.text import MIMEText
from email.mime.base import MIMEBase

def alert(main_msg):
   mail_from = 'antonyhyson@protonmail.com'
   mail_to = 'antonyhyson@gmail.com'
   msg = MIMEMultipart()
   msg['From'] = mail_from
   msg['To'] = mail_to
   msg['Subject'] = '!Alert Mail On Product Shortage! - Regards'
   mail_body = main_msg
   msg.attach(MIMEText(mail_body))
   print(msg)

   try:
      server = smtplib.SMTP_SSL('smtp.sendgrid.net', 465)
      server.ehlo()
      server.login('Flask-mail',os.environ.get('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'))
      server.sendmail(mail_from, mail_to, msg.as_string())
      server.close()
      print("Mail sent successfully!")
   except:
      print("Some Issue, Mail not Sent :(")