Final Deliverable/Final Code - Inventory Managment System for Retailers/Backend/requirements.txt
Flask
ibm_db
sendgrid
python-dotenv
flask_cors
flask_mail
secure-smtplib